refactor: 드랍다운 컴포넌트 확장 및 폴더 구조 변경 (#127)
